issue 42 : ensure runtime type of variable definition is kept

+++ SCT2/trunk/test-plugins/org.yakindu.sct.simulation.core.sexec.test/src/org/yakindu/sct/model/sexec/interpreter/test/util/AbstractExecutionFlowTest.java
@@ -60,9 +60,9 @@ public abstract class AbstractExecutionFlowTest {
 		return (Boolean) variable.getValue();
 	}
 
-	protected Float getReal(String varName) {
+	protected Double getReal(String varName) {
 		ExecutionVariable variable = context().getVariable(varName);
-		return (Float) variable.getValue();
+		return (Double) variable.getValue();
 	}
 
 	protected String getString(String varName) {
